pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "i3-notifier"
version = "0.23"
description = "A notification daemon for i3"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Sencer Selcuk" }
]
dependencies = [
    "pyxdg==0.26",
    "python-daemon==2.2.4",
    "i3ipc"
]

[project.scripts]
i3-notifier = "i3notifier.main:main"
switch-to-urgent = "i3notifier.switch_to_urgent:main"

[tool.hatch.build.targets.wheel]
packages = ["i3notifier"]

[tool.hatch.build.targets.sdist]
include = [
    "i3notifier",
    "tests",
    "LICENSE",
    "README.md",
    "MANIFEST.in",
    "examples",
    "images"
]

[tool.ruff]
indent-width = 2

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true

